fixing compiler warnings

+++ lib/lock/yang_lock.cpp
@@ -74,7 +74,7 @@ public:
         Fence();
         if (rival != -1)
         { // there is someone competing
-            if (*tiebreaker == thread_id)
+            if (*tiebreaker == static_cast<int>(thread_id))
             { // this thread_id either set the tiebreaker after the rival, or the rival has yet to set
                 Fence();
                 if (spinners[rival-starting_thread_id] == 0)
@@ -94,7 +94,7 @@ public:
                 } // wait until rival either says they updated tiebreaker or they have finished crit section
                 Fence();
 
-                if (*tiebreaker == thread_id)
+                if (*tiebreaker == static_cast<int>(thread_id))
                 { // we were later in setting tiebreaker
                     while (spinners[thread_id-starting_thread_id] !=2)
                     {
@@ -121,7 +121,7 @@ public:
         Fence();
         int rival = *tiebreaker;           // find out if you have a rival
 
-        if (rival != thread_id)
+        if (rival != static_cast<int>(thread_id))
         {                        // you have a competitor who is waiting
             spinners[rival-starting_thread_id] = 2; // free the competitor
             Fence();
@@ -148,7 +148,7 @@ public:
 private:
     int side(size_t thread_id)
     {
-        if (thread_id < starting_thread_id + (int)(num_threads / 2))
+        if (thread_id < starting_thread_id + (num_threads / 2))
         {
             return 0;
         }
